+++ CHANGELOG.md
@@ -3,7 +3,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### 🐞 Fixed
+- Reaction overlay display in a modal chat view
 
 # [4.30.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.30.0)
 _March 30, 2023_